requirements.txt
flask
flask-cors
python-dotenv
deepgram-sdk
boto3
requests
